% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotop_3d.R
\name{isotop_3d}
\alias{isotop_3d}
\title{Create an HTML interactive chart with the values of 3 different isotops}
\usage{
isotop_3d(
  df = NA,
  vars = c("var1", "var2", "var3"),
  marker.size = 5,
  symbols = "circle",
  color = NA,
  colors = NA,
  background.color = "black",
  title.plot = "LIA measurements",
  export.plot = FALSE,
  out.plot = "isotop_3d.html",
  dirOut = paste0(system.file(package = "itineRis"), "/results/")
)
}
\arguments{
\item{df}{the dataframe. The color is displayed on the 'object' field}

\item{vars}{the names of the columns where the isotops measurments are}

\item{background.color}{color of the plotly background, 'black' or 'white',
by default 'black'}

\item{title.plot}{the title of the plot}

\item{export.plot}{if TRUE, export the plot}

\item{out.plot}{the name of the output 3D plot}

\item{dirOut}{path of the output directory}

\item{color.column}{conventionaly it is 'color', dot, and the name of the field
used to get the colors. By default 'object' ('color.object')}
}
\value{
a plotly HML 3D plot
}
\description{
get a dataframe and create a 3D plot
}
\examples{

df.isotop <- isotop_dataframe()
isotop_3d(df.isotop)


}
