% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotop_ca.R
\name{isotop_ca}
\alias{isotop_ca}
\title{Correspondance Analysis on isotop data}
\usage{
isotop_ca(
 
    df.iso = "https://raw.githubusercontent.com/zoometh/itineRis/main/inst/extdata/example_1.csv",
  iso.preselect = NA,
  num_column = "num",
  col.group = NA,
  color.def.ind = "blue",
  color.def.var = "black",
  pt_siz = 1.5,
  lbl.size = 2,
  print.caption = TRUE,
  export.plot = FALSE,
  dirOut = paste0(system.file(package = "itineRis"), "/results/"),
  ca.name = "ca.png",
  plot.width = 22.2,
  plot.height = 19.5,
  plot.dpi = 300,
  verbose = TRUE
)
}
\arguments{
\item{df.iso}{a dataframe with isotop results}

\item{iso.preselect}{a preselection of the isotops that will be analysed, in a vector form, ex: `c(1, 2, 3, 4, 5)`. If provided, will skip the user prompt. Default NA.}

\item{num_column}{the column name of assemblage numbers}

\item{col.group}{the column name of the sample group (a site, a culture, etc.). Default: NA. Will only affect the symbol color by overwriting the `color.def.ind` variable.}

\item{pt_siz}{the size of the symbols}

\item{print.caption}{if TRUE will print information in the console (group colors, etc.)}

\item{export.plot}{if TRUE (Default: FALSE), save the plot. If not, will plot it}

\item{dirOut}{the output directory}

\item{ca.name}{the name of the output plot if saved}

\item{plot.width, plot.height, plot.dpi}{the dimensions and resolution of the plot if saved}

\item{verbose}{if TRUE (default): verbose}
}
\value{
a gglot with the different CA by period
}
\description{
Correspondance Analysis on isotop data. Inspired by the `zoo_ca()` function of zoowork
}
\examples{

# Plot, with colors on groups, Itineris ANR data
isotop_ca(col.group = "Provenienza",
          pt_siz = 2,
          lbl.size = 3)

# Run on preselected isotops (1 to 5) only, Shadreck Chirikure data
isotop_ca(df.iso = "C:/Users/Thomas Huet/Desktop/Shadreck data/XRF_Majors.csv",
          iso.preselect = c(1, 2, 3, 4, 5))

# Export
isotop_ca(export.plot = T,
          dirOut = "C:/Users/Thomas Huet/Desktop/Shadreck data/")

}
