% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotop_dataframe.R
\name{isotop_dataframe}
\alias{isotop_dataframe}
\title{Create a simple dataframe with colors depending on a field}
\usage{
isotop_dataframe(
  df = NA,
  df.path = paste0(system.file(package = "itineRis"), "/extdata/isotop_results.tsv"),
  color.column = "object",
  verbose = TRUE
)
}
\arguments{
\item{df}{a dataframe. By defaut NA, so it will read a file using `df.path`}

\item{df.path}{if `df` is NA, will use this path and name to read the dataset from a file. Either a TSV or a CSV with ';' separators.}

\item{color.column}{the name on which the different colors will be setup. By default, the colors
are those of the objects}
}
\value{
a dataframe with an hexadecimal value for the colors
}
\description{
Creates a new colum for the color based on another column. For example 'color.objects'
means the colors are based on the 'object' field (selected by default)
}
\examples{

df.isotop <- isotop_dataframe()

}
