% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotop_datatable.R
\name{isotop_datatable}
\alias{isotop_datatable}
\title{Create an HTML interactive table  with the values of 3 different isotops}
\usage{
isotop_datatable(
  df = NA,
  color.column = "object",
  out.plot = "isotop_datatable.html",
  dirOut = paste0(system.file(package = "itineRis"), "/results/")
)
}
\arguments{
\item{df}{the dataframe. The color is displayed on the 'object' field}

\item{color.column}{the name on which the different colors will be setup. By default,
the 'object' column}

\item{out.plot}{the path and name of the created intearctive data table}
}
\value{
an DT table with a confirmation message
}
\description{
get a simple dataframe and create an interactive datatable.
 This function will reuse the same colors as 'isotop_dataframe()' so it has
  to be run after this function
}
\examples{

# color by default, on 'object' column
df.isotop <- isotop_dataframe()
isotop_datatable(df.isotop)

# color on 'site' column
df.isotop <- isotop_dataframe(color.column = 'site')
isotop_datatable(df.isotop, color.column = 'site')

}
