% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotop_thesaurus.R
\name{isotop_thesaurus}
\alias{isotop_thesaurus}
\title{Read the thesaurus of authoritative labels for chemistry variables}
\usage{
isotop_thesaurus(
 
    path.thes = "https://raw.githubusercontent.com/zoometh/itineRis/main/inst/extdata/isotops_thesaurus.csv",
  as.list = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path.thes}{the path to the thesaurus file.}

\item{as.list}{if TRUE (Default: FALSE) return all values, and colnames, as a vector. If FALSE, return the thesaurus as a dataframe.}

\item{verbose}{if TRUE, print messages}
}
\value{
A vector of values
}
\description{
Read the thesaurus of isotopes, elements, ... Add values and column names into a list that can be compared to input dataset.
}
\examples{

# Return the thesaurus as a vector of values
thes <- isotop_thesaurus(as.list = T)

}
