% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_leaflet.R
\name{map_leaflet}
\alias{map_leaflet}
\title{Create an HTML interactive map for the site distribution}
\usage{
map_leaflet(
  d = NA,
  d.coords = NA,
  df.path = NA,
  site_column = "site",
  x_column = "x",
  y_column = "y",
  color.point = "red",
  radius.point = 5,
  export.plot = F,
  out.plot = "map_sites.html",
  dirOut = paste0(system.file(package = "itineRis"), "/results/")
)
}
\arguments{
\item{d}{a hash dictionnary-like object to store the data. By default NA}

\item{df.path}{the path to the dataframe. This dataframe must have at least three columns: one for the site name (by defaut 'site'), two for the coordinates (by default, 'x' and 'y'). By default NA, then will read the `d` variable}

\item{export.plot}{if TRUE, export the map}

\item{out.plot}{the name of the output map}

\item{dirOut}{path of the output directory}
}
\value{
a leaflet interactive map
}
\description{

}
\examples{

d <- hash::hash()
d <- read_mines(d = d,
                df.path = "C:/Rprojects/itineRis/results/Coordinates-mines_Thomas Huet.xlsx")
map_leaflet(d = d, d.coords = "mines.coords", export.plot = T)


map_leaflet(df.path = paste0(system.file(package = "itineRis"), "/extdata/sites_coords.tsv"),
            color.point = "blue")

}
