% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coords.R
\name{read_coords}
\alias{read_coords}
\title{Read the dataset of mines}
\usage{
read_coords(
  d = NA,
  df.path = paste0(system.file(package = "itineRis"),
    "/extdata/Coordinates-mines_Thomas Huet.xlsx"),
  df.sheetnames = c("all"),
  extract.mines = TRUE,
  pattern.mines.name = "locality",
  pattern.mines.coords = c("longitude", "latitude")
)
}
\arguments{
\item{d}{a hash dictionnary-like object to store the data}

\item{df.path}{the path and name of the dataset. An XLSX file.}

\item{df.sheetnames}{the name or index of the spreadsheet(s) to be read. By default "all" will read all the spreadsheets}

\item{extract.mines}{if TRUE, will extract the coordinates of one to many spreadsheets and store them under the "mines.coords" key}

\item{pattern.mines.coords}{if `extract.mines` is TRUE, will search (regex) for these patterns. A vector of two strings: first longitude/X, second: Y/latitude. By default `c("longitude", "latitude")`}
}
\value{
A hash
}
\description{
Read the dataset of mines. Will store every spreadsheet under its name within the hash object (e.g. "France")
}
\examples{

d <- hash::hash()
d <- read_coords(d = d,
                df.path = "C:/Rprojects/itineRis/results/Coordinates-mines_Thomas Huet.xlsx")

}
