% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lia.R
\name{read_lia}
\alias{read_lia}
\title{Read XLSX dataset of lead isotop analysis (LIA)}
\usage{
read_lia(
  d = NA,
  d.tag = "lia",
  df.path = paste0(system.file(package = "itineRis"), "/extdata/LIA data objects.xlsx"),
  df.sheetnames = c("all"),
  header.line = 1,
  pattern.objects.num = "Labornr",
  pattern.objects.Pb206_Pb204 = "Pb206/Pb204",
  pattern.objects.Pb207_Pb204 = "Pb207/Pb204",
  pattern.objects.Pb208_Pb204 = "Pb208/Pb204",
  export.dataframe = FALSE,
  out.df = "lia.tsv",
  dirOut = paste0(system.file(package = "itineRis"), "/results/")
)
}
\arguments{
\item{d}{a hash dictionnary-like object to store the data}

\item{d.tag}{the new field in `d`. By default, 'lia'}

\item{df.path}{the path and name of the dataset. An XLSX file.}

\item{df.sheetnames}{the name or index of the spreadsheet(s) to be read. By default "all" will read all the spreadsheets}

\item{header.line}{the row of the headers, by default the first line}

\item{pattern.objects.num}{Search (regex) for this pattern in the column names to find the laboratory number. It is important to remind that R doesn't accept spaces in header that must be replcaed by dots}

\item{pattern.objects.Pb206_Pb204, pattern.objects.Pb207_Pb204, pattern.objects.Pb208_Pb204}{Search (regex) for these patterns in the column names to find the LIA measurements}

\item{dirOut}{path of the output directory}

\item{export.plot}{if TRUE, export the dataframe}

\item{out.plot}{the name of the output dataframe}
}
\value{
A hash with a new entrance. If export is TRUE, will export as TSV
}
\description{
Read the dataset of objects. Will store every spreadsheet under its name within the hash object (e.g. "golasecca")
}
\examples{

d <- hash::hash()
d <- read_lia(d = d,
              d.tag = "lia.objects",
              df.path = "C:/Rprojects/itineRis/results/LIA data objects.xlsx",
              header.line = 2)
d <- read_lia(d = d,
              d.tag = "lia.mines",
              pattern.objects.num = "Locality/.Mine",
              pattern.objects.Pb206_Pb204 = "206Pb/204Pb",
              pattern.objects.Pb207_Pb204 = "207Pb/204Pb",
              pattern.objects.Pb208_Pb204 = "208Pb/204Pb",
              df.path = "C:/Rprojects/itineRis/results/Coordinates-mines_Thomas Huet.xlsx",
              header.line = 1)

}
