% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesaurus.R
\name{thesaurus}
\alias{thesaurus}
\title{Read the Itineris thesaurus}
\usage{
thesaurus(
  inDir = "https://raw.githubusercontent.com/ANR-Itineris/itineris/main/lod/thesaurus",
  inFile = "Itineris_th101.rdf",
  mod = "read",
  root = "artefact",
  prefix = "https://thesaurus.mom.fr:443/opentheso/?idc=",
  suffix = "&idt=th101",
  outPlot = "collapsibleTree",
  export.plot = FALSE,
  outFile = NA,
  outDir = system.file(package = "itineRis"),
  verbose = FALSE
)
}
\arguments{
\item{mod}{the way the thesaurus in handled. Default, "read".}

\item{root}{the node under which the subgraph will be filtered}

\item{prefix, suffix}{the prefix and suffix to remove.}

\item{outPlot}{the type of interactive plot.}

\item{export.plot}{if TRUE will export into outDir.}

\item{outDir}{the output directory.}

\item{verbose}{if TRUE, print messages}

\item{path.thes}{the path to the thesaurus file.}
}
\value{
an interactive plot
}
\description{
Read the multi-linguism thesaurus hosted by https://thesaurus.mom.fr/opentheso/?idt=th101 and replot it as an interactive graph.
}
\examples{

## export the subgraph of 'artefact' in a collapsibleTree layout
thesaurus(root = "artefact",
          outPlot = "collapsibleTree",
          export.plot = T,
          outDir = "C:/Rprojects/itineRis/results/")


}
