% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_leaflet.R
\name{map_leaflet}
\alias{map_leaflet}
\title{Create an HTML interactive map for the site distribution}
\usage{
map_leaflet(
  df.path = paste0(system.file(package = "itineRis"), "/extdata/sites_coords.tsv"),
  site_column = "site",
  x_column = "x",
  y_column = "y",
  color.point = "red",
  radius.point = 5,
  out.plot = "map_sites.html",
  dirOut = paste0(system.file(package = "itineRis"), "/results/")
)
}
\arguments{
\item{df.path}{the path to the dataframe. This dataframe must have at least three columns:
one for the site name (by defaut 'site'), two for the coordinates (by default, 'x' and 'y')}

\item{out.plot}{the path and name of the created map}
}
\value{
a leaflet interactive map
}
\description{

}
\examples{

map_leaflet(color.point = "blue")

}
